python-service/requirements.txt
flake8
pytest
coverage
pandas
sqlalchemy
pymysql
requests
python-dotenv